+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Adjustment: replace logging/trace_on_debug by new log level "trace"
 * Adjustment: sharing/token: adjust default permissions to "rp"
 * Fix: sharing/propfind+proppatch: permission check related to properties
+* Fix: hook/email: add support for date-only events
+* Feature: hook/email: in case of mass-email was enabled but only one attendee fall-back to non-mass-email
 
 ## 3.7.0